Fixed double promotion.

+++ flight/PiOS/Common/pios_mpxv5004.c
@@ -93,7 +93,7 @@ float PIOS_MPXV5004_ReadAirspeed(uint8_t airspeedADCPin)
 	}
 	
 	//Compute calibraterd airspeed, as per http://en.wikipedia.org/wiki/Calibrated_airspeed
-	float calibratedAirspeed = STANDARD_AIR_MACH_SPEED*sqrtf(5.0f*(pow(Qc/(STANDARD_AIR_SEA_LEVEL_PRESSURE/1000)+1.0f,POWER)-1.0f));
+	float calibratedAirspeed = STANDARD_AIR_MACH_SPEED*sqrtf(5.0f*(powf(Qc/(STANDARD_AIR_SEA_LEVEL_PRESSURE/1000)+1.0f,POWER)-1.0f));
 	
 	//Upper bound airspeed. No need to lower bound it, that comes from Qc
 	if (calibratedAirspeed > 80) { //in [m/s]
